+++ src/analyze/engine/index.ts
@@ -5,19 +5,30 @@ import chalk from "chalk";
 import requestEngine from "./requestEngine.js";
 import esqueryEngine from "./esqueryEngine.js";
 
-export const engine = async (rule: Rule, mappedJsonData: Chunks | undefined, openapiData: OpenAPISpec | undefined) => {
+export const engine = async (
+    rule: Rule,
+    mappedJsonData: Chunks | undefined,
+    openapiData: OpenAPISpec | undefined,
+    tech: string
+) => {
     // first of all check what is rule type, and then check if the data for that is available or is undefined
 
     if (rule.type === "request") {
         if (!openapiData) {
             return;
         }
-        requestEngine(rule, openapiData);
+
+        if (tech.split(",").includes(rule.tech) || tech === "all") {
+            requestEngine(rule, openapiData);
+        }
     } else if (rule.type === "esquery") {
         if (!mappedJsonData) {
             return;
         }
-        esqueryEngine(rule, mappedJsonData);
+
+        if (tech.split(",").includes(rule.tech) || tech === "all") {
+            esqueryEngine(rule, mappedJsonData);
+        }
     }
 };